Change nai column to smallint

+++ backend/aggregated_history/helper_functions.sql
@@ -5,7 +5,7 @@ SET ROLE btracker_owner;
 DROP TYPE IF EXISTS btracker_backend.balance_history_return CASCADE;
 CREATE TYPE btracker_backend.balance_history_return AS (
     account     INT,
-    nai         INT,
+    nai         SMALLINT,
     balance     BIGINT,
     min_balance BIGINT,
     max_balance BIGINT,
